+++ src/prefect/serialization/environment.py
@@ -1,4 +1,5 @@
-from marshmallow import fields
+from marshmallow import fields, post_load
+from typing import Any
 
 from prefect.environments import (
     DaskKubernetesEnvironment,
@@ -6,7 +7,7 @@ from prefect.environments import (
     LocalEnvironment,
     RemoteEnvironment,
 )
-from prefect.utilities.serialization import ObjectSchema, OneOfSchema
+from prefect.utilities.serialization import ObjectSchema, OneOfSchema, to_qualified_name
 
 
 class BaseEnvironmentSchema(ObjectSchema):
@@ -33,10 +34,6 @@ class DaskKubernetesEnvironmentSchema(ObjectSchema):
     min_workers = fields.Int()
     max_workers = fields.Int()
 
-    # Serialize original spec file locations, not the spec itself
-    scheduler_spec_file = fields.String(allow_none=True)
-    worker_spec_file = fields.String(allow_none=True)
-
 
 class RemoteEnvironmentSchema(ObjectSchema):
     class Meta:
@@ -45,6 +42,26 @@ class RemoteEnvironmentSchema(ObjectSchema):
     executor = fields.String(allow_none=True)
     executor_kwargs = fields.Dict(allow_none=True)
     labels = fields.List(fields.String())
+
+
+class CustomEnvironmentSchema(ObjectSchema):
+    class Meta:
+        object_class = lambda: Environment
+        exclude_fields = ["type"]
+
+    labels = fields.List(fields.String())
+
+    type = fields.Function(
+        lambda environment: to_qualified_name(type(environment)), lambda x: x
+    )
+
+    @post_load
+    def create_object(self, data: dict, **kwargs: Any) -> Environment:
+        """
+        Because we cannot deserialize a custom class, we return an empty
+        Base Environment with the appropriate labels.
+        """
+        return Environment(labels=data.get("labels"))
 
 
 class EnvironmentSchema(OneOfSchema):
@@ -58,4 +75,12 @@ class EnvironmentSchema(OneOfSchema):
         "Environment": BaseEnvironmentSchema,
         "LocalEnvironment": LocalEnvironmentSchema,
         "RemoteEnvironment": RemoteEnvironmentSchema,
+        "CustomEnvironment": CustomEnvironmentSchema,
     }
+
+    def get_obj_type(self, obj: Any) -> str:
+        name = obj.__class__.__name__
+        if name in self.type_schemas:
+            return name
+        else:
+            return "CustomEnvironment"